+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feature: install déjà dup backups
 - feature: install draw.io
+- feature: install dropbox
 
 ## [1.4.0] - 2026-07-24